nptl: Extract <bits/atomic_wide_counter.h> from pthread_cond_common.c
And make it an installed header.  This addresses a few aliasing
violations (which do not seem to result in miscompilation due to
the use of atomics), and also enables use of wide counters in other
parts of the library.

The debug output in nptl/tst-cond22 has been adjusted to print
the 32-bit values instead because it avoids a big-endian/little-endian
difference.

Reviewed-by: Adhemerval Zanella  <adhemerval.zanella@linaro.org>
(cherry picked from commit 8bd336a00a5311bf7a9e99b3b0e9f01ff5faa74b)

Conflicts:
	nptl/Makefile
		Adapt to different overall layout
	nptl/tst-cond22.c
		Updated for context due to
		c36fc50781995e6758cae2b6927839d0157f213c
		being already backported
	sysdeps/nptl/bits/thread-shared-types.h
		Updated for context due to
		c36fc50781995e6758cae2b6927839d0157f213c
		being already backported