elf: Add _dl_find_object function
It can be used to speed up the libgcc unwinder, and the internal
_dl_find_dso_for_object function (which is used for caller
identification in dlopen and related functions, and in dladdr).

_dl_find_object is in the internal namespace due to bug 28503.
If libgcc switches to _dl_find_object, this namespace issue will
be fixed.  It is located in libc for two reasons: it is necessary
to forward the call to the static libc after static dlopen, and
there is a link ordering issue with -static-libgcc and libgcc_eh.a
because libc.so is not a linker script that includes ld.so in the
glibc build tree (so that GCC's internal -lc after libgcc_eh.a does
not pick up ld.so).

It is necessary to do the i386 customization in the
sysdeps/x86/bits/dl_find_object.h header shared with x86-64 because
otherwise, multilib installations are broken.

The implementation uses software transactional memory, as suggested
by Torvald Riegel.  Two copies of the supporting data structures are
used, also achieving full async-signal-safety.

Reviewed-by: Adhemerval Zanella  <adhemerval.zanella@linaro.org>
(cherry picked from commit 5d28a8962dcb6ec056b81d730e3c6fb57185a210)

Conflicts:
	elf/Makefile
		Updated for change of layout
	elf/dl-support.c
		Updated for minor context changes
	elf/rtld.c
		Updated for minor context changes
	include/link.h
		Updated for minor context changes
	manual/dynlink.texi
		Already added by glibc-RHEL-22165-1.patch, folded in the
		typo fix from upstream
		6cf4ebe10c6f0f60cdfce98f5a0ec7c5ceb987df
	sysdeps/mach/hurd/i386/libc.abilist
		Updated for minor context changes