glibc-rh2162962.patch
Maintain an explicit order of tunables, so that the tunable_list
array and the tunable_id_t constant can remain consistent over time.

Related to this upstream bug:

  Internal tunables ABI depends on awk array iteration order
  <https://sourceware.org/bugzilla/show_bug.cgi?id=30027>

The new dl-tunables.list files are already on the sysdeps search
path, which is why the existing Makeconfig rule picks them up.
The files for RHEL 9.1z were created by applying the gen-tunables.awk
part of this patch to RHEL 9.1.0 (glibc-2.34-40.el9_2.1, to be
precise).  The sysdeps/unix/sysv/linux/**/dl-tunables.list files were
created based on the generated error message during the RHEL 9.1.z
build.

Going forward, new tunables will have to be added manually to the end
of those files.  Existing tunables should not be deleted.  For
deletion, the script would have to be extended to be able to create
gaps in the tunable_list array.