glibc-RHEL-55471-9.patch
Downstream-only patch to restore the extern ABI for functions
like fprintf that use the FILE * type.  Rebuilds of applications
receive ABI change reports because of this installed header change
(indirect subtype change in libabigail terms), and given that
this part of struct _IO_FILE is strictly internal, there is no
need to expose this change to installed headers.